requirements.txt
fastapi==0.121.3
langchain==1.0.8
langchain_community==0.4.1
langchain_core==1.0.6
langchain_openai==1.0.3
langgraph==1.0.3
loguru==0.7.3
pydantic==2.12.4
python-dotenv==1.2.1
regex==2022.7.9
regex==2024.11.6
regex==2023.10.3
Requests==2.32.5
uvicorn==0.38.0
